Remove libpq type aliases from global namespace. (#1190)
Fixes: #1189.

In the process, I changed the mechanism by which I defined these type aliases.  They are now aliases for `void`, not undefined structs.  It doesn't hurt since all that matters is that the pointer's type as accessed matches the object's original type.  There was no need to declare C linkage, or to put them in the global namespace.

The change to `void` also made it possible to replace `reinterpret_cast` with `static_cast` in a few places.  The cost is a whole bunch of casting.  I also changed the type of `params`' formats array to match the libpq interpretation more closely, which also let me replace some of those casts.

+++ src/params.cxx
@@ -119,7 +119,7 @@ pqxx::internal::c_params pqxx::params::make_c_params(sl loc) const
             check_cast<int>(std::ssize(value), s_overflow, loc));
         }
 
-        p.formats.push_back(param_format(value));
+        p.formats.push_back(static_cast<int>(param_format(value)));
       },
       param);
